requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
django-phonenumber-field==7.0.2
django-taggit==3.1.0
djangorestframework==3.14.0
Markdown==3.4.1
phonenumbers==8.13.7
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
